Debugging the code for generating the VSBGraph from the voxels.
Added a new file to the regression suite for testing the VSBGraph
and its clipping operations.
The All2x2x2 test generates all 256 configurations of voxels
in a 2x2x2 cube to check that the graph edges are connected
correctly.  This version is segfaulting on the checkerboard
configuration.  I'm committing this now to move it to Linux to run
valgrind.

Added a --noclean option to regression.py, to keep it from removing its
temp directory.

+++ TEST3D/regression.py
@@ -108,11 +108,11 @@ if __name__=="__main__":
     try:
         opts,args = getopt.getopt(sys.argv[1:],"f:a:t:o:d",
                                   ["from=", "after=", "to=", "oofargs=",
-                                   "forever", "debug", "backwards",
+                                   "forever", "debug", "backwards", "noclean",
                                    "dryrun"])
     except getopt.GetoptError, err:
         print str(err)
-        print "Usage: regression.py [--from <starttest> | --after <starttest>] [--to <endtest>] [--forever] [--oofargs <oofargs>] [--debug] [--backwards] [tests]"
+        print "Usage: regression.py [--from <starttest> | --after <starttest>] [--to <endtest>] [--forever] [--oofargs <oofargs>] [--debug] [--noclean] [--backwards] [tests]"
         print "       Don't use --from or --to if tests are listed explicitly."
         sys.exit(2)
 
@@ -123,6 +123,7 @@ if __name__=="__main__":
     forever = False
     debug = False
     backwards = False
+    noclean = False
 
     global dryrun
     dryrun = False
@@ -158,6 +159,8 @@ if __name__=="__main__":
             debug = True
         elif o == "--backwards":
             backwards = True
+        elif o == "--noclean":
+            noclean = True
         elif o in ("-d", "--dryrun"):
             dryrun = True
 
@@ -223,8 +226,10 @@ if __name__=="__main__":
         OOF.File.Quit()
     finally:
         if ok:
-            os.rmdir(tmpdir)
             print >> sys.stderr, "All tests completed successfully!"
+            if not noclean:
+                os.rmdir(tmpdir)
         else:
             print >> sys.stderr, "Tests failed."
+        if noclean or not ok:
             print >> sys.stderr, "Temp dir", tmpdir, "was not removed."
